pyenv_requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
anyio==4.12.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.2
cachetools==6.2.4
catboost==1.2.8
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
cmake==4.2.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==4.5.0
decorator==5.2.1
dill==0.4.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Farama-Notifications==0.0.4
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.12.19
folium==0.20.0
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
graphviz==0.21
grpcio==1.76.0
gymnasium==1.1.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
keras-facenet==0.3.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightgbm==4.6.0
llvmlite==0.43.0
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mtcnn==0.1.1
multidict==6.7.0
multiprocess==0.70.18
namex==0.1.0
narwhals==2.14.0
networkx==3.2.1
nltk==3.9.2
numba==0.60.0
numpy==2.0.2
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
platformdirs==4.4.0
plotly==6.5.0
polars==1.36.1
polars-runtime-32==1.36.1
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytorch-tabnet==4.1.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==5.1.2
sentencepiece==0.1.99
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
stable_baselines3==2.7.1
streamlit==1.50.0
streamlit-folium==0.25.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.22
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
torch-geometric==2.6.1
torchvision==0.23.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.5
typer-slim==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.7
ultralytics-thop==2.0.18
Unidecode==1.4.0
urllib3==2.6.2
watchdog==6.0.0
Werkzeug==3.1.4
wordcloud==1.9.5
wrapt==2.0.1
xgboost==2.1.4
xxhash==3.6.0
xyzservices==2025.11.0
yarl==1.22.0
zipp==3.23.0
zxcvbn==4.5.0
